refactor: 유틸함수 timeSince 분리
